+++ src/main/java/controller/exception/ProgramQuitException.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.exception;
 
 public class ProgramQuitException extends RuntimeException {
     public static final String MESSAGE = "프로그램을 종료합니다.";
