+++ src/main/java/reader/exception/FileReadException.java
@@ -1,4 +1,4 @@
-package reader;
+package reader.exception;
 
 import java.io.IOException;
